+++ examples/object_geos_parts.rs
@@ -9,7 +9,7 @@ fn main() -> Result<()> {
 
     let lib = session.load_asset_file("otls/sesi/SideFX_spaceship.hda")?;
     let node = lib.try_create_first()?;
-    node.cook_blocking(None)?;
+    node.cook_blocking()?;
     let _asset_info = node.asset_info()?;
 
     for info in node.get_objects_info()? {