+++ examples/packed_primitives.rs
@@ -23,12 +23,12 @@ fn main() -> Result<()> {
             }
         );
         co.set_packed_prim_instancing_mode(mode);
-        asset.cook_blocking(Some(&co))?;
+        asset.cook_with_options(&co, true)?;
 
         let nodes = asset.find_children_by_type(NodeType::Sop, NodeFlags::Any, false)?;
         for handle in nodes {
             let node = handle.to_node(&session)?;
-            node.cook_blocking(Some(&co))?;
+            node.cook_with_options(&co, true)?;
             let geo = node.geometry()?.expect("geometry");
             println!(
                 "Part count for node {:?}: {}",