+++ src/geometry.rs
@@ -69,12 +69,18 @@ impl Geometry {
     }
 
     pub fn part_info(&self, part_id: i32) -> Result<PartInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_part_info(&self.node, part_id).map(|inner| PartInfo { inner })
     }
 
     pub fn volume_info(&self, part_id: i32) -> Result<VolumeInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_volume_info(&self.node, part_id).map(|inner| VolumeInfo { inner })
     }
 
@@ -84,12 +90,18 @@ impl Geometry {
     }
 
     pub fn volume_bounds(&self, part_id: i32) -> Result<VolumeBounds> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_volume_bounds(&self.node, part_id)
     }
 
     pub fn geo_info(&self) -> Result<GeoInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         GeoInfo::from_node(&self.node)
     }
 
@@ -99,13 +111,19 @@ impl Geometry {
     }
 
     pub fn box_info(&self, part_id: i32) -> Result<BoxInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_box_info(self.node.handle, &self.node.session, part_id)
             .map(|inner| BoxInfo { inner })
     }
 
     pub fn sphere_info(&self, part_id: i32) -> Result<BoxInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_box_info(self.node.handle, &self.node.session, part_id)
             .map(|inner| BoxInfo { inner })
     }
@@ -166,25 +184,37 @@ impl Geometry {
     }
 
     pub fn curve_info(&self, part_id: i32) -> Result<CurveInfo> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_curve_info(&self.node, part_id).map(|inner| CurveInfo { inner })
     }
 
     /// Retrieve the number of vertices for each curve in the part.
     pub fn curve_counts(&self, part_id: i32, start: i32, length: i32) -> Result<Vec<i32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_curve_counts(&self.node, part_id, start, length)
     }
 
     /// Retrieve the orders for each curve in the part if the curve has varying order.
     pub fn curve_orders(&self, part_id: i32, start: i32, length: i32) -> Result<Vec<i32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_curve_orders(&self.node, part_id, start, length)
     }
 
     /// Retrieve the knots of the curves in this part.
     pub fn curve_knots(&self, part_id: i32, start: i32, length: i32) -> Result<Vec<f32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_curve_knots(&self.node, part_id, start, length)
     }
 
@@ -192,7 +222,10 @@ impl Geometry {
     /// ith element in the array is the point index the ith vertex
     /// associates with.
     pub fn vertex_list(&self, part: Option<&PartInfo>) -> Result<Vec<i32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part, self.part_info(0)?);
         crate::ffi::get_geo_vertex_list(
@@ -205,14 +238,20 @@ impl Geometry {
     }
 
     pub fn partitions(&self) -> Result<Vec<PartInfo>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         (0..self.info.part_count())
             .map(|i| self.part_info(i))
             .collect()
     }
 
     pub fn get_face_counts(&self, part: Option<&PartInfo>) -> Result<Vec<i32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part, self.part_info(0)?);
         crate::ffi::get_face_counts(
@@ -225,7 +264,10 @@ impl Geometry {
     }
 
     pub fn get_materials(&self, part: Option<&PartInfo>) -> Result<Option<Materials>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part, self.part_info(0)?);
         let (all_the_same, mats) = crate::ffi::get_material_node_ids_on_faces(
@@ -261,7 +303,10 @@ impl Geometry {
     }
 
     pub fn get_group_names(&self, group_type: GroupType) -> Result<StringArray> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let count = match group_type {
             GroupType::Point => self.info.point_group_count(),
             GroupType::Prim => self.info.primitive_group_count(),
@@ -272,7 +317,10 @@ impl Geometry {
     }
 
     pub fn get_edge_count_of_edge_group(&self, group: &str, part_id: i32) -> Result<i32> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let group = CString::new(group)?;
         crate::ffi::get_edge_count_of_edge_group(
             &self.node.session,
@@ -287,7 +335,10 @@ impl Geometry {
         owner: AttributeOwner,
         part: Option<&PartInfo>,
     ) -> Result<StringArray> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part, self.part_info(0)?);
         let counts = part.attribute_counts();
@@ -303,6 +354,10 @@ impl Geometry {
     }
 
     pub fn get_position_attribute(&self, part_id: i32) -> Result<NumericAttr<f32>> {
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let name = CString::new("P")?;
         let inner = crate::ffi::get_attribute_info(
             &self.node,
@@ -323,7 +378,10 @@ impl Geometry {
         owner: AttributeOwner,
         name: &str,
     ) -> Result<Option<Attribute>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let name = CString::new(name)?;
         let inner = crate::ffi::get_attribute_info(&self.node, part_id, owner, &name)?;
         let storage = inner.storage;
@@ -390,7 +448,10 @@ impl Geometry {
         part_id: i32,
         info: AttributeInfo,
     ) -> Result<StringAttr> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let name = CString::new(name)?;
         crate::ffi::add_attribute(&self.node, part_id, &name, &info.inner)?;
         Ok(StringAttr::new(name, info, self.node.clone()))
@@ -479,7 +540,10 @@ impl Geometry {
         group_type: GroupType,
         group_name: &str,
     ) -> Result<Vec<i32>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part_info, self.part_info(0)?);
         let group_name = CString::new(group_name)?;
@@ -498,13 +562,20 @@ impl Geometry {
         group_type: GroupType,
         info: Option<&GeoInfo>,
     ) -> Result<i32> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let info = unwrap_or_create!(tmp, info, self.geo_info()?);
         Ok(crate::ffi::get_group_count_by_type(info, group_type))
     }
 
     pub fn get_instanced_part_ids(&self, part_info: Option<&PartInfo>) -> Result<Vec<i32>> {
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part_info, self.part_info(0)?);
         crate::ffi::get_instanced_part_ids(
@@ -519,7 +590,10 @@ impl Geometry {
         &self,
         part_info: Option<&PartInfo>,
     ) -> Result<(i32, i32)> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part_info, self.part_info(0)?);
         crate::ffi::get_group_count_on_instance_part(
@@ -534,7 +608,10 @@ impl Geometry {
         group: GroupType,
         part_id: i32,
     ) -> Result<StringArray> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         crate::ffi::get_group_names_on_instance_part(
             &self.node.session,
             self.node.handle,
@@ -548,7 +625,10 @@ impl Geometry {
         part_info: Option<&PartInfo>,
         order: RSTOrder,
     ) -> Result<Vec<Transform>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tmp;
         let part = unwrap_or_create!(tmp, part_info, self.part_info(0)?);
         crate::ffi::get_instanced_part_transforms(
@@ -562,7 +642,10 @@ impl Geometry {
     }
 
     pub fn save_to_file(&self, filepath: &str) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let path = CString::new(filepath)?;
         crate::ffi::save_geo_to_file(&self.node, &path)
     }
@@ -584,7 +667,10 @@ impl Geometry {
     }
 
     pub fn save_to_memory(&self, format: GeoFormat) -> Result<Vec<i8>> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let format = unsafe { CStr::from_bytes_with_nul_unchecked(format.as_c_literal()) };
         crate::ffi::save_geo_to_memory(&self.node.session, self.node.handle, format)
     }
@@ -602,7 +688,10 @@ impl Geometry {
         tile: &VolumeTileInfo,
         values: &mut [T],
     ) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         T::read_tile(&self.node, part, fill, values, &tile.inner)
     }
 
@@ -612,7 +701,10 @@ impl Geometry {
         tile: &VolumeTileInfo,
         values: &[T],
     ) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         T::write_tile(&self.node, part, values, &tile.inner)
     }
 
@@ -624,7 +716,10 @@ impl Geometry {
         z_index: i32,
         values: &mut [T],
     ) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         T::read_voxel(&self.node, part, x_index, y_index, z_index, values)
     }
 
@@ -636,7 +731,10 @@ impl Geometry {
         z_index: i32,
         values: &[T],
     ) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         T::write_voxel(&self.node, part, x_index, y_index, z_index, values)
     }
 
@@ -646,7 +744,10 @@ impl Geometry {
         info: &VolumeInfo,
         callback: impl Fn(Tile),
     ) -> Result<()> {
-        debug_assert!(self.node.is_valid()?);
+        debug_assert!(
+            self.node.get_info()?.total_cook_count() > 0,
+            "Node not cooked"
+        );
         let tile_size = (info.tile_size().pow(3) * info.tuple_size()) as usize;
         crate::volume::iterate_tiles(&self.node, part, tile_size, callback)
     }
@@ -744,12 +845,12 @@ mod tests {
         geo.set_vertex_list(0, [0, 1, 2]).unwrap();
         geo.set_face_counts(0, [3]).unwrap();
         geo.commit().expect("commit");
-        geo.node.cook_blocking(None).unwrap();
+        geo.node.cook_blocking().unwrap();
     }
 
     fn _load_test_geometry(session: &Session) -> super::Result<Geometry> {
         let node = session.create_node("Object/hapi_geo", None, None)?;
-        node.cook(None).unwrap();
+        node.cook().unwrap();
         node.geometry()
             .map(|some| some.expect("must have geometry"))
     }
@@ -867,7 +968,7 @@ mod tests {
             let geo = session.create_input_node("dummy").unwrap();
             geo.load_from_file(&tmp_file.to_string_lossy())
                 .expect("load_from_file");
-            geo.node.cook(None).unwrap();
+            geo.node.cook().unwrap();
             assert_eq!(geo.part_info(0).unwrap().point_count(), 3);
             geo.node.delete().unwrap();
         });
@@ -898,10 +999,10 @@ mod tests {
             let geo = session.create_input_node("input").unwrap();
             _create_triangle(&geo);
             geo.commit().unwrap();
-            geo.node.cook_blocking(None).unwrap();
+            geo.node.cook_blocking().unwrap();
             assert_eq!(geo.part_info(0).unwrap().point_count(), 3);
             geo.revert().unwrap();
-            geo.node.cook_blocking(None).unwrap();
+            geo.node.cook_blocking().unwrap();
             assert_eq!(geo.part_info(0).unwrap().point_count(), 0);
             geo.node.delete().unwrap();
         });
@@ -915,7 +1016,7 @@ mod tests {
             geo.add_group(0, GroupType::Point, "test", Some(&[1, 1, 1]))
                 .unwrap();
             geo.commit().unwrap();
-            geo.node.cook_blocking(None).unwrap();
+            geo.node.cook_blocking().unwrap();
             assert_eq!(
                 geo.group_count_by_type(GroupType::Point, geo.geo_info().as_ref().ok()),
                 Ok(1)
@@ -923,7 +1024,7 @@ mod tests {
 
             geo.delete_group(0, GroupType::Point, "test").unwrap();
             geo.commit().unwrap();
-            geo.node.cook_blocking(None).unwrap();
+            geo.node.cook_blocking().unwrap();
             assert_eq!(geo.group_count_by_type(GroupType::Point, None), Ok(0));
             geo.node.delete().unwrap();
         });
@@ -935,7 +1036,7 @@ mod tests {
             let node = session.create_node("Object/hapi_geo", None, None).unwrap();
             let opt = CookOptions::default()
                 .with_packed_prim_instancing_mode(PackedPrimInstancingMode::Flat);
-            node.cook_blocking(Some(&opt)).unwrap();
+            node.cook_with_options(&opt, true).unwrap();
             let outputs = node.geometry_outputs().unwrap();
             let instancer = outputs.get(1).unwrap();
             let ids = instancer.get_instanced_part_ids(None).unwrap();
@@ -960,7 +1061,7 @@ mod tests {
     fn get_face_materials() {
         with_session(|session| {
             let node = session.create_node("Object/spaceship", None, None).unwrap();
-            node.cook(None).unwrap();
+            node.cook().unwrap();
             let geo = node.geometry().expect("geometry").unwrap();
             let mats = geo.get_materials(None).expect("materials");
             assert!(matches!(mats, Some(Materials::Single(_))));
@@ -983,12 +1084,12 @@ mod tests {
     fn read_write_volume() {
         with_session(|session| {
             let node = session.create_node("Object/hapi_vol", None, None).unwrap();
-            node.cook_blocking(None).unwrap();
+            node.cook_blocking().unwrap();
             let source = node.geometry().unwrap().unwrap();
             let source_part = source.part_info(0).unwrap();
             let vol_info = source.volume_info(0).unwrap();
             let dest_geo = session.create_input_node("volume_copy").unwrap();
-            dest_geo.node.cook_blocking(None).unwrap();
+            dest_geo.node.cook_blocking().unwrap();
             dest_geo.set_part_info(&source_part).unwrap();
             dest_geo.set_volume_info(0, &vol_info).unwrap();
 
@@ -1004,7 +1105,7 @@ mod tests {
                 })
                 .unwrap();
             dest_geo.commit().unwrap();
-            dest_geo.node.cook_blocking(None).unwrap();
+            dest_geo.node.cook_blocking().unwrap();
         });
     }
 }