+++ src/lib.rs
@@ -103,6 +103,8 @@
 //! which return CString or String. See the [`stringhandle`] module for more info.
 //!
 
+extern crate core;
+
 pub mod asset;
 pub mod attribute;
 pub mod geometry;