+++ src/material.rs
@@ -136,7 +136,7 @@ mod tests {
     fn extract_image() {
         with_session(|session| {
             let node = session.create_node("Object/spaceship", None, None).unwrap();
-            node.cook(None).unwrap();
+            node.cook().unwrap();
             let geo = node.geometry().expect("geometry").unwrap();
             let mats = geo.get_materials(None).expect("materials");
             if let Some(Materials::Single(mat)) = mats {